requirements.txt
altgraph==0.17     
amqp==5.0.1    
appdirs==1.4.4    
asgiref==3.2.10   
auto-py-to-exe==2.7.8    
beautifulsoup4==4.9.2    
billiard ==3.6.3.0  
bottle==0.12.18  
bottle-websocket==0.2.9    
CacheControl==0.12.6   
celery==5.0.0    
certifi==2020.6.20
chardet==3.0.4    
click==7.1.2    
click-didyoumean==0.0.3    
click-repl==0.1.6    
colorama==0.4.3    
contextlib2==0.6.0    
coreapi==2.3.3    
coreschema==0.0.4    
defusedxml==0.6.0    
distlib==0.3.1    
distro==1.4.0    
Django==3.1.1    
django-allauth==0.42.0   
django-anysign==1.4      
django-cors-headers==3.5.0    
django-crispy-forms==1.9.2    
django-crontab==0.7.1    
django-docusign==3.3      
django-filter==2.3.0    
django-jquery==3.1.0    
django-phone-field==1.8.1    
django-phonenumber-field==5.0.0    
django-rest-auth==0.9.5    
django-rest-swagger==2.1.2    
django-session-timeout==0.1.0    
django-static-jquery==2.1.4    
django-storages==1.10.1   
djangorestframework==3.11.1   
djangorestframework-jwt==1.11.0   
djangorestframework-simplejwt==4.4.0    
drf-yasg==1.17.1   
Eel==0.12.4   
filelock==3.0.12   
  
PyMySQL==0.10.1   
pyparsing==2.4.7    
python-barcode==0.13.1   
python-dateutil==2.8.1    
python-decouple==3.3      
python3-openid==3.2.0    
pytoml==0.1.21   
pytz==2020.1   
requests==2.24.0   
requests-oauthlib==1.3.0    
retrying==1.3.3    
ruamel.yaml==0.16.12  
ruamel.yaml.clib==0.2.2    
setuptools==50.3.1   
simplejson==3.17.2   
six==1.15.0   
soupsieve==2.0.1    
SQLAlchemy==1.3.19   
sqlparse==0.3.1    
uritemplate==3.0.1    
urllib3==1.25.10  
vine==5.0.0    
virtualenv==20.0.33  
virtualenv-clone==0.5.4    
wcwidth==0.2.5    
webencodings==0.5.1    
Werkzeug==1.0.1    
wheel==0.34.2   
whichcraft==0.6.1    
whitenoise==5.2.0    
xlwt==1.3.0    
zipp==3.2.0    
zope.event==4.5.0    
zope.interface==5.1.0

